:sparkles: Adding a `Pulsar` model.
This commit adds a `models` module to pataka, and also adds a `Pulsar` model to it. This is a class that helps the user simulate a pulsar, without going into the nitty-gritty of the core API. The `make` method, that creates an instance of this class, is packed with sensible defaults.