:bug: Fixing the Gaussian pulse profile.
This commit fixes a bug that I found in the Gaussian pulse profile I was making. The profile being made was neither of the width, nor the amplitude specified by the user. Using the FWHM for the width, and scaling the profile by its maxima before scaling it by the amplitude, fixed the bug.